+++ src/Products/MeetingBEP/config.py
@@ -18,9 +18,6 @@ DEFAULT_ADD_CONTENT_PERMISSION = "Add portal content"
 setDefaultRoles(DEFAULT_ADD_CONTENT_PERMISSION, ('Manager', 'Owner', 'Contributor'))
 product_globals = globals()
 
-# "Human ressources (Confidential)" group id
-HR_CONFIDENTIAL_GROUP_ID = 'ressources-humaines-confidentiel'
-
 # "DU" auto replaced content values
 DU_ORIGINAL_VALUE = "<p><strong><u>Proposition de décision&nbsp;:</u></strong></p>"
 DU_RATIFICATION_VALUE = "<p><u><strong>Le {mc_title} décide à l'unanimité de ratifier la décision " \